requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
requests
aiohttp